+++ CHANGELOG.md
@@ -19,13 +19,15 @@ and this project attempts to adhere to [Semantic Versioning](https://semver.org/
   - Customize the max size of the returned base64 image with the `maxImageSizeMB` option.
 - New `getBulkMetadata` function.
   - Slightly different than `getMetadata` as this function won't throw an error. It'll return an object with 2 fields, `results` & `errors` which contain an array of objects containing the URI & data/error associated with it.
+- New `saveArtwork` function.
+  - Allows the user to save the embedded artwork to a file and optionally compress it to save on space.
 
 ### ⚙️ Internal Changes
 
 - Reorganization of code.
   - Now "normalize" and put all the field-value pairs inside a map, which we then send the ones we want to return.
 - Ensure we release `MetadataRetriever` & `MediaMetadataRetriever` when we're done using them.
-- Updated example app to use `getBulkMetadata` instead of `getMetadata`.
+- Updated example app to use `getBulkMetadata` & `saveArtwork` instead of `getMetadata`.
 
 ## [1.1.0] - 2025-08-21